move IUserDataProfileService to workbench (#152427)
