+++ src/vs/platform/storage/browser/storageService.ts
@@ -13,7 +13,7 @@ import { assertIsDefined } from 'vs/base/common/types';
 import { InMemoryStorageDatabase, isStorageItemsChangeEvent, IStorage, IStorageDatabase, IStorageItemsChangeEvent, IUpdateRequest, Storage } from 'vs/base/parts/storage/common/storage';
 import { ILogService } from 'vs/platform/log/common/log';
 import { AbstractStorageService, IS_NEW_KEY, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
-import { IUserDataProfile, IUserDataProfileService } from 'vs/platform/userDataProfile/common/userDataProfile';
+import { IUserDataProfile } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { IAnyWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
 
 export class BrowserStorageService extends AbstractStorageService {
@@ -42,12 +42,12 @@ export class BrowserStorageService extends AbstractStorageService {
 
 	constructor(
 		private readonly payload: IAnyWorkspaceIdentifier,
+		currentProfile: IUserDataProfile,
 		@ILogService private readonly logService: ILogService,
-		@IUserDataProfileService userDataProfileService: IUserDataProfileService
 	) {
 		super({ flushInterval: BrowserStorageService.BROWSER_DEFAULT_FLUSH_INTERVAL });
 
-		this.globalStorageProfile = userDataProfileService.currentProfile;
+		this.globalStorageProfile = currentProfile;
 	}
 
 	private getId(scope: StorageScope): string {
